Move billboard_line to rviz2 for use in grid
